requirements.txt
pandas==1.0.4
tqdm==4.48.2
bert4keras==0.10.6
numpy==1.19.4
tensorflow_gpu==1.14.0
Keras==2.3.1
scikit_learn==1.0.1
